Update transform imports with restructuring
